streamline file system setup and enhance script explorer fixtures
